Add metrics@custom component template to metrics-*-* index template
This lets users customize the metrics data stream mappings, without having to override a managed component template that may get overridden.
